+++ libs/react-components/src/tedi/components/label/label.spec.tsx
@@ -1,4 +1,4 @@
-import { render } from '@testing-library/react';
+import { fireEvent, render } from '@testing-library/react';
 
 import { useBreakpointProps } from '../../helpers';
 import { Label } from './label';
@@ -77,5 +77,37 @@ describe('Label component', () => {
     const { container } = render(<Label required>Label</Label>);
     const required = container.querySelector('.tedi-label__required');
     expect(required).toBeInTheDocument();
+  });
+
+  it('renders InfoButton when infoButtonProps are provided', () => {
+    const mockOnClick = jest.fn();
+
+    const { container, getByRole } = render(
+      <Label
+        infoButton={{
+          onClick: mockOnClick,
+          children: 'More Info',
+          title: 'Additional Info',
+        }}
+        required
+      >
+        Label
+      </Label>
+    );
+
+    const infoButton = getByRole('button');
+    expect(infoButton).toBeInTheDocument();
+    const required = container.querySelector('.tedi-label__required');
+    expect(required?.nextSibling).toBe(infoButton);
+
+    fireEvent.click(infoButton);
+    expect(mockOnClick).toHaveBeenCalledTimes(1);
+  });
+
+  it('does not render InfoButton if infoButtonProps is not provided', () => {
+    const { queryByRole } = render(<Label>Label</Label>);
+
+    const infoButton = queryByRole('button', { name: 'More Info' });
+    expect(infoButton).not.toBeInTheDocument();
   });
 });