+++ libs/react-components/src/tedi/index.ts
@@ -13,13 +13,19 @@ export * from './components/status-badge/status-badge';
 export * from './components/separator/separator';
 export * from './components/buttons/closing-button/closing-button';
 export * from './components/print/print';
+export * from './components/form/textfield/textfield';
+export * from './components/form/textarea/textarea';
+export * from './components/form/number-field/number-field';
+export * from './components/form/feedback-text/feedback-text';
 export * from './components/buttons/button/button';
 export * from './components/buttons/info-button/info-button';
 export * from './components/notifications/alert/alert';
 export * from './components/notifications/toast/toast';
 export * from './components/stretch-content/stretch-content';
 export * from './components/navigation/link/link';
-export * from './components/form/feedback-text/feedback-text';
+export * from './components/section/section';
+export * from './components/hide-on-scroll/hide-on-scroll';
+export * from './components/truncate/truncate';
 
 export * from './helpers';